Remove unnecessary code from runtime API messages
